Delete unused code
The syntax VAR=${VAR:-`command`} is to provide a default value for VAR
in case is not defined. In this case, the default value is being taken
from SSM on AWS, but we are not using that service. We are providing
those values as environment variables on Porter configuration so there
is no need to have that code in the file.